Have Travis use k3s to test the containers.
Test basic Pulp functionality, including writing to disk Using
pulp_file's sync tests.

Also, don't set CONTENT_HOST in the pulp-server ConfigMap.
If set incorrectly, it causes pulp_file test scripts to fail.
We haven't implemented querying it via the K8s APIs,
or having users specify it yet.

Note that the RequestsDependencyWarning messages from httpie are
annoying output, but pip3 is not on Ubuntu on Travis.
It appears that Travis relies on the virtual env for Python3

fixes #5061
https://pulp.plan.io/issues/5061