switch to postgres 9.6 centos7 image from docker.io rather than registry.centos.org
This fixes k3s being unable to download it from registry.centos.org,
which is due to either the fact that it requires the docker:// v1 protocol,
or it being an insecure registry.

It seems to be the same image, but newer by 1 month & a small diff in
size.
Both by CentOS themselves. Same args.

We cannot use the RHEL8 postgres image because it is publicly searchable
but not publicly accessible.

Note that an attempt was tried to use docker://registry.centos.org as the image
name, but it randomly stopped working (STATUS: InvalidImageName)