Problem: pulp user can't create database for unit tests
Solution: use 'postgres' admin user instead

The database connection username and password can now be configured using a resource.
The username, password, and amin_password can be provided.

When admin_password is provided, it is used to set the 'postgres' user's password.
This admin user can then be used to run create a test database for running unit tests.

The username and password are used to create a user account that has access to 'pulp'
database. This account is then used to connect all the services to the database.