k3s-install.sh: Problem: The node info or storage commands fail
depending on timing.

Due to the node not being up by the time systemd says the service started.

Solution: Use bash functionality to wait for it the socket to be listening.
Then wait for cluster info & node status commands to succeed.

re: #5373
As a user, I can use a single script to install k3s and launch pulp-operator
https://pulp.plan.io/issues/5375