Remove deployment_state variable
The default behavior of an Ansible Operator is to delete all resources
the operator created during reconciliation when a managed resource is
marked for deletion.

Because of the above statement, having a deployment_state variable
becomes not needed.

Also, has a sanity for reconciliation loop, let's use `apply = true` to
ensure if the resource already exist it is re-applied nonetheless to be
always in the state desired.

Source: https://v0-19-x.sdk.operatorframework.io/docs/ansible/reference/finalizers/

[noissue]