Redis: Fix Permission denied and multiple deployment in same namespace
This PR fixes the following issues that were happening in the redis pod

```
24 Mar 2021 10:21:16.062 # Failed opening the RDB file dump.rdb (in
server root dir /data) for saving: Permission denied
24 Mar 2021 10:21:16.162 # Background saving error
```

As it is written today. PVC is always made with `redis-data` name and
does not take in consideration the deployment name. This allows for
multiple deployment of pulp in single namespace.

[noissue]