Fix bug on db controller and "improve" ci test
The CI script is not checking if the API pod is READY to serve the
requests. It is only checking if the pod is running.
As a workaround this commit adds a "grep" to search the
"Listening" string on API logs to consider it as READY to serve
requests.

(cherry picked from commit a3a9094b679f62dc12ddc09cc1719d19b492c95d)