Remove networkpolicy from default deploy step
The allow-same-namespace network policy allows the communication only
for pods running in the same namespace. For clusters installed with
ingress_type = { ingress | route } the communication from ingress
controller pods to pulp pods will be denied. Considering that we are
not managing the network policies through the operator controller nor
through OLM/CSV we are removing this config, so that running a
`make deploy` would be more similar to what OLM delivers.
https://issues.redhat.com/browse/AAP-5673
[noissue]

Co-authored-by: Christian M. Adams <chadams@redhat.com>