Fix "spamming" retry in case of error
From what I could understand, during an error condition we were
calling the Status.Update() method in every retry of loop reconcile (even
without anything needing to be changed).
This was kind of breaking the idempotency, so each time we called the
Status.Update() method I think the controller thought that it was a
new iteration, which caused the exponential backoff retry to not increment.

I'm not 100% sure if this will fix all of the "missing" exponential backoff retry.
fixes #650