Fix tests based on .status.conditions update
Most of our envtest tests were relying on a "fake" status.
envtest does not have a running k8s controller, so the operator
`.status.conditions` field would never get into a READY state because
they are updated based on the `.status` of another k8s controller (the deployment).
This commit modifies the logic on how some tests are made to check
if a field has been modified by controller according to the expected.
[noissue]