Remove reconcile trigger for some specific errors
Some errors, like a missing/wrong configuration will never be fixed
by triggering a reconciliation. For such errors, the controller will
keep spamming error messages (with an exponential backoff retry) until
the problem is fixed, which is also consuming resources without need.
This commit is removing the reconciliation request for some errors
that would not be fixed by retriggering the reconcile method.
ref: #822
[noissue]