Add reconciliation for old postgres svc
Golang version is keeping the database service deployed by ansible to
avoid having to touch the secrets (which could have been manually modified
by users). In this case, we need to also watch for any modification in this
service with the golang controller because if users modify or delete it they
will be lost on how to proceed and pulp-api pods would lose access to the
database.

[noissue]