Forcing specific image version in galaxy samples
Momentarily hardcoding image version for galaxy samples to avoid breaking
pipeline execution. We had some issues with the image itself modification
and also with the galaxy source update.
To avoid more problems that are hard to troubleshoot (using latest/nightly tags
we had to check if the image and/or the source code was/were modified) this
commit is temporarily fixing an image version until we find a better approach.
[noissue]