Temporarily pin the pulp-web version in CI samples
pulp-oci-image repo has been updated and we are currently missing
the amd64 architecture for pulp-web image in Quay. As a workaround,
this commit pins to an older image that still supports amd64.
It also updates the check for docker installation in CI because
the ubuntu 24.04 (GH runner) does not come with the daemon.json
file by default.

[noissue]